Linux 3.14 with MCST patches

+++ drivers/net/ethernet/intel/e1000/e1000_hw.c
@@ -5085,8 +5085,8 @@ void e1000_get_bus_info(struct e1000_hw *hw)
  */
 static void e1000_write_reg_io(struct e1000_hw *hw, u32 offset, u32 value)
 {
-	unsigned long io_addr = hw->io_base;
-	unsigned long io_data = hw->io_base + 4;
+	unsigned long io_addr = (unsigned long)hw->io_base;
+	unsigned long io_data = (unsigned long)hw->io_base + 4;
 
 	e1000_io_write(hw, io_addr, offset);
 	e1000_io_write(hw, io_data, value);
